chore(ymax-planner): Move deprecated `handleDeposit` function
...into the test file that needs it
